Move ProfileService to UserService
