+++ LiveLines.Api/Users/IUserService.cs
@@ -7,4 +7,6 @@ public interface IUserService
     Task<LoggedInUser> CreateUser(string provider, string username);
     Task<LoggedInUser> GetUser(string username);
     Task<LoggedInUser> GetUser(int userId);
+    Task<Profile> UpdateProfile(LoggedInUser user, ProfileToUpdate profileToUpdate);
+    Task<Profile> GetProfile(LoggedInUser user);
 }
