+++ LiveLines.Users/UserService.cs
@@ -6,15 +6,20 @@ namespace LiveLines.Users;
 public class UserService : IUserService
 {
     private readonly IUserStore _userStore;
+    private readonly IProfileStore _profileStore;
 
-    public UserService(IUserStore userStore)
+    public UserService(IUserStore userStore, IProfileStore profileStore)
     {
         _userStore = userStore;
+        _profileStore = profileStore;
     }
         
     public async Task<LoggedInUser> CreateUser(string provider, string username)
     {
-        return await _userStore.CreateUser(provider, username);
+        LoggedInUser user = await _userStore.CreateUser(provider, username);
+        await _profileStore.CreateProfile(user);
+        
+        return user;
     }
 
     public async Task<LoggedInUser> GetUser(string username)
@@ -25,5 +30,15 @@ public class UserService : IUserService
     public async Task<LoggedInUser> GetUser(int userId)
     {
         return await _userStore.GetUser(userId);
+    }
+
+    public async Task<Profile> UpdateProfile(LoggedInUser user, ProfileToUpdate profileToUpdate)
+    {
+        return await _profileStore.UpdateProfile(user, profileToUpdate.LinePrivacy);
+    }
+
+    public async Task<Profile> GetProfile(LoggedInUser user)
+    {
+        return await _profileStore.GetProfile(user);
     }
 }
